refactor: 회원가입 타입 모두 type 디렉토리로 변경 (#44)
