+++ src/components/register/FormValidation.tsx
@@ -1,64 +1,4 @@
-// 인증 유효성 검사 규칙
-export const AUTH_VALIDATION = {
-	PASSWORD: {
-		MIN_LENGTH: 6,
-	},
-} as const;
-
-// 역할 타입과 옵션
-export type Role = 'ticket' | 'snack' | 'floor';
-export const ROLE_OPTIONS = {
-	ticket: { value: 'ticket', label: '매표' },
-	snack: { value: 'snack', label: '매점' },
-	floor: { value: 'floor', label: '플로어' },
-} as const;
-
-// 직책 타입과 옵션
-export type Position = 'admin' | 'user';
-export const POSITION_OPTIONS = {
-	admin: { value: 'admin', label: '관리자' },
-	user: { value: 'user', label: '사용자' },
-} as const;
-
-// 성별 타입과 옵션
-export type Gender = '남성' | '여성';
-export const GENDER_OPTIONS = ['남성', '여성'] as const;
-
-// FormData 타입 정의
-export type FormData = {
-	email: string;
-	password: string;
-	confirmPassword: string;
-	role: Role | '';
-	gender: Gender | '';
-	age: string;
-	position: Position | '';
-	workingHours: string;
-};
-
-// FormErrors 타입 정의
-export type FormErrors = {
-	[K in keyof FormData]?: string;
-};
-
-// 에러 메시지 상수
-export const ERROR_MESSAGES = {
-	REQUIRED: {
-		email: '이메일을 입력해주세요',
-		password: '비밀번호를 입력해주세요',
-		confirmPassword: '비밀번호 확인을 입력해주세요',
-		role: '역할을 선택해주세요',
-		gender: '성별을 선택해주세요',
-		age: '나이를 입력해주세요',
-		position: '직책을 입력해주세요',
-	},
-	INVALID: {
-		email: '올바른 이메일 형식이 아닙니다',
-		password: `비밀번호는 ${AUTH_VALIDATION.PASSWORD.MIN_LENGTH}자 이상이어야 합니다`,
-		confirmPassword: '비밀번호가 일치하지 않습니다',
-		age: '올바른 나이를 입력해주세요',
-	},
-} as const;
+import { FormData, FormErrors, ERROR_MESSAGES, AUTH_VALIDATION } from '@/pages';
 
 // 폼 유효성 검사 함수
 export const validateForm = (data: FormData): FormErrors => {
