+++ src/pages/Home/Home.tsx
@@ -2,12 +2,9 @@ import { useEffect, useState } from 'react';
 import { db } from '@/firebaseConfig';
 import { doc, setDoc } from 'firebase/firestore';
 import * as S from './Home.styles';
-import { Button } from '@/components';
-import CheckboxGroup from '@/components/checkbox/CheckboxGroup';
 import { Button, Error, Loading } from '@/components';
 import ModalPortal from '@/components/modal/ModalPortal';
 import ScheduleModal from '@/components/modal/ScheduleModal';
-
 
 export function Home() {
 	const [isOpen, setIsOpen] = useState<boolean>(false);
